requirements.txt
openai>=1.0.0
ollama>=0.1.0
faiss-cpu>=1.7.4
numpy>=1.24.0
python-dotenv>=1.0.0
PyPDF2>=3.0.0
pydantic>=2.0.0
rich>=13.0.0          # CLI formatting
langchain>=0.1.0      # Optional: for advanced RAG
tiktoken>=0.5.0       # Token counting